refactor(structure): move components into a directory inside the home page dir
- so files are better organized and easier to find
