+++ app/components/Cover.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 
-import heroIllustrationMobile from "../public/images/hero-illustration-mobile.svg";
-import heroIllustrationDesktop from "../public/images/hero-illustration-desktop.svg";
+import heroIllustrationMobile from "../../public/images/hero-illustration-mobile.svg";
+import heroIllustrationDesktop from "../../public/images/hero-illustration-desktop.svg";
 import ContactUs from "./ContactUs";
 
 export default function Cover() {
