+++ app/components/Stack.tsx
@@ -1,31 +1,31 @@
 import Image from "next/image";
 
-import rubyOnRails from "../public/images/logos/rails-logo.svg";
-import sidekiq from "../public/images/logos/sidekiq-logo.svg";
-import rubocop from "../public/images/logos/rubocop-logo.svg";
-import rspec from "../public/images/logos/rspec-logo.svg";
-import react from "../public/images/logos/react-logo.svg";
-import mui from "../public/images/logos/mui-logo.svg";
-import nextjs from "../public/images/logos/nextjs-logo.svg";
-import typescript from "../public/images/logos/typescript-logo.svg";
-import tailwindcss from "../public/images/logos/tailwindcss-logo.svg";
-import prettier from "../public/images/logos/prettier-logo.svg";
-import eslint from "../public/images/logos/eslint-logo.svg";
-import heroku from "../public/images/logos/heroku-logo.svg";
-import vercel from "../public/images/logos/vercel-logo.svg";
-import posgres from "../public/images/logos/posgres-logo.svg";
-import redis from "../public/images/logos/redis-logo.svg";
-import reactNative from "../public/images/logos/react-native-logo.svg";
-import expo from "../public/images/logos/expo-logo.svg";
-import github from "../public/images/logos/github-logo.svg";
-import nodejs from "../public/images/logos/nodejs-logo.svg";
-import nestjs from "../public/images/logos/nestjs-logo.svg";
-import graphql from "../public/images/logos/graphql-logo.svg";
-import aws from "../public/images/logos/aws-logo.svg";
-import serverless from "../public/images/logos/serverless-logo.svg";
-import gem from "../public/images/gem.svg";
-import paperCoffeeCup from "../public/images/paper-coffee-cup.svg";
-import Logo from "./Logo";
+import rubyOnRails from "../../public/images/logos/rails-logo.svg";
+import sidekiq from "../../public/images/logos/sidekiq-logo.svg";
+import rubocop from "../../public/images/logos/rubocop-logo.svg";
+import rspec from "../../public/images/logos/rspec-logo.svg";
+import react from "../../public/images/logos/react-logo.svg";
+import mui from "../../public/images/logos/mui-logo.svg";
+import nextjs from "../../public/images/logos/nextjs-logo.svg";
+import typescript from "../../public/images/logos/typescript-logo.svg";
+import tailwindcss from "../../public/images/logos/tailwindcss-logo.svg";
+import prettier from "../../public/images/logos/prettier-logo.svg";
+import eslint from "../../public/images/logos/eslint-logo.svg";
+import heroku from "../../public/images/logos/heroku-logo.svg";
+import vercel from "../../public/images/logos/vercel-logo.svg";
+import posgres from "../../public/images/logos/posgres-logo.svg";
+import redis from "../../public/images/logos/redis-logo.svg";
+import reactNative from "../../public/images/logos/react-native-logo.svg";
+import expo from "../../public/images/logos/expo-logo.svg";
+import github from "../../public/images/logos/github-logo.svg";
+import nodejs from "../../public/images/logos/nodejs-logo.svg";
+import nestjs from "../../public/images/logos/nestjs-logo.svg";
+import graphql from "../../public/images/logos/graphql-logo.svg";
+import aws from "../../public/images/logos/aws-logo.svg";
+import serverless from "../../public/images/logos/serverless-logo.svg";
+import gem from "../../public/images/gem.svg";
+import paperCoffeeCup from "../../public/images/paper-coffee-cup.svg";
+import Logo from "../../components/Logo";
 import StackGroup from "./StackGroup";
 
 export default function Stack() {
