+++ app/components/StackGroup.tsx
