+++ app/components/Values.tsx
@@ -1,8 +1,8 @@
 import Image from "next/image";
 
-import singingFrogWithCloudDesktop from "../public/images/singing-frog-with-cloud-desktop.svg";
-import singingFrogWithCloudMobile from "../public/images/singing-frog-with-cloud-mobile.svg";
-import underline from "../public/images/underline.svg";
+import singingFrogWithCloudDesktop from "../../public/images/singing-frog-with-cloud-desktop.svg";
+import singingFrogWithCloudMobile from "../../public/images/singing-frog-with-cloud-mobile.svg";
+import underline from "../../public/images/underline.svg";
 import Value from "./Value";
 
 export default function Values() {
